Fixed slicing

+++ src/twomartens/masterthesis/aae/train.py
@@ -171,7 +171,7 @@ def _train_one_epoch_simple(epoch: int,
             enc_dec_loss_avg(reconstruction_loss)
             
             if int(global_step % LOG_FREQUENCY) == 0:
-                comparison = K.concatenate([x[:batch_size/2], x_decoded[:batch_size/2]], axis=0)
+                comparison = K.concatenate([x[:int(batch_size / 2)], x_decoded[:int(batch_size / 2)]], axis=0)
                 grid = util.prepare_image(comparison.cpu(), nrow=int(batch_size/2))
                 summary_ops_v2.image(name='reconstruction',
                                      tensor=K.expand_dims(grid, axis=0), max_images=1,
